# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Solar Project Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List scan history
      responses:
        "200":
          description: Scan history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanRun"
    post:
      operationId: triggerScan
      tags: [scans]
      summary: Trigger a new scan of all sources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "202":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanRun"
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan run status and results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanRun"
        "404":
          description: Not found
  /scans/{id}/projects:
    get:
      operationId: getScanProjects
      tags: [scans]
      summary: List all projects discovered by a scan
      description: Returns every qualifying project linked to this scan, including new projects, dated updates to canonical projects, and current inventory observations.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of projects with isNew flag
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanProject"
        "404":
          description: Scan not found
  /scans/{id}/sources:
    get:
      operationId: getScanSources
      tags: [scans]
      summary: List durable source-acquisition health for a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: One safe acquisition outcome per attempted approved source
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanSourceHealth"
        "404":
          description: Scan not found
...
components:
  schemas:
...
    ScanRun:
      type: object
      required: [id, status, startedAt, sourcesScanned, projectsFound]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [running, completed, failed]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        sourcesScanned:
          type: integer
        projectsFound:
          type: integer
        newProjects:
          type: integer
        errorMessage:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
    ScanInput:
      type: object
      properties:
        startDate:
          type: string
          description: Only scan for projects announced after this date (YYYY-MM-DD)
        endDate:
          type: string
          description: Only scan for projects announced before this date (YYYY-MM-DD)
    ScanProject:
      type: object
      required: [id, name, country, status, createdAt, updatedAt, isNew, eventType, effectiveDate, dateEvidence, eventSourceUrl, eventSourceName]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        capacityMw:
          type: ["number", "null"]
        developer:
          type: ["string", "null"]
        epc:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        country:
          type: string
          enum: [AU, NZ]
        status:
          type: string
          enum: [announced, under_development]
        sourceUrl:
          type: ["string", "null"]
        sourceName:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        announcedDate:
          type: ["string", "null"]
        announcedDateEvidence:
          type: ["string", "null"]
        lastSeenAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        scanId:
          type: ["integer", "null"]
        isNew:
          type: boolean
          description: True if this project was newly inserted during this scan
        eventType:
          type: string
          enum: [new, updated, inventory_observed]
          description: How the canonical project relates to this scan; this does not redefine isNew.
        effectiveDate:
          type: ["string", "null"]
          format: date
          description: Date of the source event, or null for an undated inventory observation.
        dateEvidence:
          type: string
          description: Provenance classification supporting the effective date.
        eventSourceUrl:
          type: ["string", "null"]
          description: Source URL for this scan event, which may differ from the canonical project's source URL.
        eventSourceName:
          type: ["string", "null"]
    ScanSourceHealth:
      type: object
      required:
        - id
        - scanId
        - sourceName
        - acquisitionMethod
        - directAttempted
        - firecrawlAttempted
        - firecrawlSucceeded
        - apifyAttempted
        - brightDataAttempted
        - openaiNormalisationAttempted
        - openaiNormalisationSucceeded
        - fallbackUsed
        - outcome
        - candidateCount
        - qualifyingProjectCount
        - durationMs
        - firecrawlCalls
        - firecrawlPages
        - firecrawlCacheReused
        - createdAt
      properties:
        id: { type: integer }
        scanId: { type: integer }
        sourceName: { type: string }
        acquisitionMethod:
          type: string
        directAttempted: { type: boolean }
        firecrawlAttempted: { type: boolean }
        firecrawlSucceeded: { type: boolean }
        apifyAttempted: { type: boolean }
        brightDataAttempted: { type: boolean }
        openaiNormalisationAttempted: { type: boolean }
        openaiNormalisationSucceeded: { type: boolean }
        fallbackUsed: { type: boolean }
        outcome:
          type: string
          enum: [success-with-results, success-zero-results, blocked, timeout, extraction-failed, missing-credentials, rate-limited, provider-error]
        candidateCount: { type: integer }
        qualifyingProjectCount: { type: integer }
        durationMs: { type: integer }
        failureCategory:
          type: ["string", "null"]
        failureReason:
          type: ["string", "null"]
        contentFingerprint:
          type: ["string", "null"]
        firecrawlCalls: { type: integer }
        firecrawlPages: { type: integer }
        firecrawlCacheReused: { type: boolean }
        createdAt: { type: string }
...